Add license information to repository
The license information is currently shown in a footer of the published page (http://gtfs.org/best-practices/), but this isn't obvious if you're looking at the repo on Github (e.g., to "borrow" some images).

This patch adds explicit license information in the README and in the standard LICENSE.md file.