Remove Best Practices already covered by the spec (#44)
* agency.txt: remove agency_lang

Redundant with specification description: "Should be provided to help GTFS consumers choose capitalization rules and other language-specific settings for the dataset."

* stops.txt: unique stop_id

Redundant as the specification indicates stop_id must be a Unique ID

* stops.txt: consistent identifiers

Already described in "Data Publishing & General Practices" section.

* stops.txt: stop_name matching published

Already defined by the specification: "The stop_name should match the agency's rider-facing name for the location as printed on a timetable, published online, or represented on signage. For translations into other languages, use translations.txt."

* stops.txt: abbreviations

Already defined under "All Files" section

* stops.txt: mixed case

Already defined under "All Files" section

* routes.txt: agency_id "must"

Covered by the spec, and "must" requirements are not allowed in best practice (recommendation) documentation.

* trips.txt: usage of direction_id

Usage of fields is covered by the spec

* stop_times.txt: timepoint

Usage of fields is covered by the spec

* stops.txt: stop_id is an internal ID

Already covered by the spec in "Field Types" section: "An ID field value is an internal ID, not intended to be shown to riders"

* stop_times.txt: stop_headsign

Covered by the spec (duplication)

* stops.txt: stop_code

Covered by the spec (duplication)

* calendar.txt & calendar_dates.txt: recommendations

Usage recommendation already covered by the spec: "Recommended: Use calendar_dates.txt in conjunction with calendar.txt to define exceptions to the default service patterns defined in calendar.txt. If service is generally regular, with a few changes on explicit dates (for instance, to accommodate special event services, or a school schedule), this is a good approach. In this case calendar_dates.service_id is a foreign ID referencing calendar.service_id."

* fare_rules.txt: fare_rules.agency_id is not a field

Remove BP for a field that does not exist in the file

* update links and contacts